+++ js/src/integrations-manager.js
@@ -1,0 +1,22 @@
+import reactIntegration from './integrations/react';
+import reactRouterIntegration from './integrations/react-router';
+
+class IntegrationsManager {
+  constructor() {
+    this.integrations = {
+      'react-component': reactIntegration.integrationWrapper,
+      'react-router': reactRouterIntegration.integrationWrapper,
+    };
+  }
+
+  get(name) {
+    // handle missing one here;
+    return this.integrations[name];
+  }
+
+  register(name, integration) {
+    this.integrations[name] = integration;
+  }
+}
+
+export default new IntegrationsManager;