+++ lib/react_webpack_rails/view_helpers.rb
@@ -1,11 +1,9 @@
 module ReactWebpackRails
   module ViewHelpers
-    def react_element(integration_name, payload = {}, options = {}, html_options = {}, &block)
-      payload = camelize_props_key(payload) if Rails.application.config.react.camelize_props
+    def react_element(integration_name, payload = {}, html_options = {}, &block)
       data = {
         integration_name: integration_name,
-        options: options,
-        payload: (payload.is_a?(String) ? payload : payload.to_json),
+        payload: payload,
         react_element: true
       }
       html_options = html_options.merge(data: data)
@@ -13,12 +11,13 @@ module ReactWebpackRails
       content_tag(html_tag, '', html_options, &block)
     end
 
-    def react_component(name, props = {}, options = {}, html_options = {})
-      react_element('react-component', props, options.merge(name: name), html_options)
+    def react_component(name, props = {}, options = { ssr: false })
+      props = camelize_props_key(props) if Rails.application.config.react.camelize_props
+      react_element('react-component', { props: props, name: name }, options)
     end
 
     def react_router(name)
-      react_element('react-router', {}, name: name)
+      react_element('react-router', name: name)
     end
 
     private
@@ -26,7 +25,8 @@ module ReactWebpackRails
     def camelize_props_key(props)
       return props unless props.is_a?(Hash)
       props.each_with_object({}) do |h, (k, v)|
-        h[k.to_s.camelize(:lower)] = v.is_a?(Hash) ? camelize_props_key(v) : v; h
+        h[k.to_s.camelize(:lower)] = v.is_a?(Hash) ? camelize_props_key(v) : v
+        h
       end
     end
   end